refactor(ipod): Split and simplify iPod app structure
- Extract PipPlayer component to its own file (179 lines)
- Extract FullScreenPortal component to its own file (810 lines)
- Extract IpodScreen helper components to screen/ folder:
  - BatteryIndicator, Scrollbar, MenuListItem, ScrollingText, StatusDisplay
- Create shared constants.ts for translation languages, themes, timing constants
- Create types.ts for shared TypeScript interfaces
- Reduce IpodAppComponent.tsx from 3,132 to 1,468 lines (53% reduction)
- Reduce IpodScreen.tsx from 908 to 505 lines (44% reduction)
- No functionality changes - all features preserved
