+++ src/apps/ipod/components/IpodScreen.tsx
@@ -1,384 +1,33 @@
-import { useState, useRef, useEffect, useLayoutEffect } from "react";
+import { useRef, useEffect } from "react";
 import ReactPlayer from "react-player";
 import { motion, AnimatePresence } from "framer-motion";
 import { cn } from "@/lib/utils";
-import { Track } from "@/stores/useIpodStore";
 import { useAppStore } from "@/stores/useAppStore";
 import { LyricsDisplay } from "./LyricsDisplay";
-import { useLyrics } from "@/hooks/useLyrics";
 import { ActivityIndicator } from "@/components/ui/activity-indicator";
-import { LyricsAlignment, ChineseVariant, KoreanDisplay, JapaneseFurigana } from "@/types/lyrics";
 import { useTranslation } from "react-i18next";
+import {
+  BatteryIndicator,
+  Scrollbar,
+  MenuListItem,
+  ScrollingText,
+  StatusDisplay,
+} from "./screen";
+import type { IpodScreenProps } from "../types";
 
-// Minimal BatteryManager interface for browsers that expose navigator.getBattery
-interface BatteryManager {
-  charging: boolean;
-  level: number;
-  addEventListener(type: string, listener: () => void): void;
-  removeEventListener(type: string, listener: () => void): void;
-}
+// Animation variants for menu transitions
+const menuVariants = {
+  enter: (direction: "forward" | "backward") => ({
+    x: direction === "forward" ? "100%" : "-100%",
+  }),
+  center: {
+    x: 0,
+  },
+  exit: (direction: "forward" | "backward") => ({
+    x: direction === "forward" ? "-100%" : "100%",
+  }),
+};
 
-// Battery component
-function BatteryIndicator({ backlightOn }: { backlightOn: boolean }) {
-  const [batteryLevel, setBatteryLevel] = useState<number | null>(null);
-  const [isCharging, setIsCharging] = useState<boolean>(false);
-  const [animationFrame, setAnimationFrame] = useState<number>(1);
-
-  useEffect(() => {
-    // Try to get battery information (deprecated API, may not work in all browsers)
-    const getBattery = async () => {
-      try {
-          if ("getBattery" in navigator) {
-            const battery = await (
-              navigator as unknown as { getBattery: () => Promise<BatteryManager> }
-            ).getBattery();
-          setBatteryLevel(battery.level);
-          setIsCharging(battery.charging);
-
-          const updateLevel = () => setBatteryLevel(battery.level);
-          const updateCharging = () => setIsCharging(battery.charging);
-
-          battery.addEventListener("levelchange", updateLevel);
-          battery.addEventListener("chargingchange", updateCharging);
-
-          return () => {
-            battery.removeEventListener("levelchange", updateLevel);
-            battery.removeEventListener("chargingchange", updateCharging);
-          };
-        }
-      } catch {
-        // Fallback to a default level
-        setBatteryLevel(1.0); // 100% as fallback
-        setIsCharging(false);
-      }
-    };
-
-    getBattery();
-  }, []);
-
-  // Animation effect for charging
-  useEffect(() => {
-    if (!isCharging) return;
-
-    const interval = setInterval(() => {
-      setAnimationFrame((prev) => (prev % 4) + 1);
-    }, 500); // Change frame every 500ms
-
-    return () => clearInterval(interval);
-  }, [isCharging]);
-
-  // Use fallback if no battery level detected
-  const level = batteryLevel ?? 1.0;
-  const filledBars = isCharging ? animationFrame : Math.ceil(level * 4);
-
-  return (
-    <div className="flex items-center">
-      {/* Battery outline */}
-      <div className="relative w-[19px] h-[10px] border border-[#0a3667] bg-transparent">
-        {/* Battery bars */}
-        <div className="absolute inset-[1px] flex gap-[1px]">
-          {[1, 2, 3, 4].map((bar) => (
-            <div
-              key={bar}
-              className={`flex-1 h-full transition-colors duration-200 ${
-                bar <= filledBars ? "bg-[#0a3667]" : "bg-transparent"
-              }`}
-            />
-          ))}
-        </div>
-      </div>
-      {/* Battery tip */}
-      <div className="w-[2px] h-[4px] bg-[#0a3667] relative">
-        <div
-          className={`w-[2px] h-[2px] absolute top-[1px] left-[-2px] right-[0px] mx-auto ${
-            backlightOn ? "bg-[#c5e0f5]" : "bg-[#8a9da9]"
-          }`}
-        />
-      </div>
-    </div>
-  );
-}
-
-// Simplified Scrollbar component
-function Scrollbar({
-  containerRef,
-  backlightOn,
-  menuMode,
-}: {
-  containerRef:
-    | React.RefObject<HTMLDivElement | null>
-    | React.MutableRefObject<HTMLDivElement | null>;
-  backlightOn: boolean;
-  menuMode: boolean;
-}) {
-  const thumbRef = useRef<HTMLDivElement>(null);
-  const trackRef = useRef<HTMLDivElement>(null);
-
-  useLayoutEffect(() => {
-    const container = containerRef.current;
-    const thumb = thumbRef.current;
-    const track = trackRef.current;
-    if (!container || !thumb || !track || !menuMode) return;
-
-    const updateScrollbar = () => {
-      const { scrollTop, scrollHeight, clientHeight } = container;
-      const needsScrollbar = scrollHeight > clientHeight;
-
-      if (needsScrollbar) {
-        track.style.opacity = "1";
-        thumb.style.display = "block";
-
-        // Account for track's extended bounds: top-[-1px] bottom-[-2px] = +3px total height
-        const trackHeight = clientHeight + 3;
-        const thumbHeight = Math.max(
-          (clientHeight / scrollHeight) * trackHeight,
-          20
-        );
-        const maxScrollTop = scrollHeight - clientHeight;
-        const thumbMaxTop = trackHeight - thumbHeight;
-        const thumbTop =
-          maxScrollTop > 0 ? (scrollTop / maxScrollTop) * thumbMaxTop : 0;
-
-        thumb.style.height = `${thumbHeight - 4}px`;
-        thumb.style.top = `${thumbTop + 2}px`; // 1px gap at top (accounting for -1px track offset)
-      } else {
-        track.style.opacity = "0";
-        thumb.style.display = "none";
-      }
-    };
-
-    // Initial update
-    updateScrollbar();
-
-    // Update on scroll
-    container.addEventListener("scroll", updateScrollbar, { passive: true });
-
-    // Update when content changes
-    const observer = new ResizeObserver(updateScrollbar);
-    observer.observe(container);
-
-    return () => {
-      container.removeEventListener("scroll", updateScrollbar);
-      observer.disconnect();
-    };
-  }, [containerRef, menuMode]);
-
-  if (!menuMode) return null;
-
-  return (
-    <div className="absolute right-0 top-[-1px] bottom-[-2px] w-2 z-20">
-      {/* Track - visibility controlled by opacity */}
-      <div
-        ref={trackRef}
-        className={cn(
-          "w-full h-full border border-[#0a3667] transition-all duration-500",
-          backlightOn
-            ? "bg-[#c5e0f5] bg-gradient-to-b from-[#d1e8fa] to-[#e0f0fc]"
-            : "bg-[#8a9da9]"
-        )}
-        style={{ opacity: 0 }}
-      />
-      {/* Thumb - visibility controlled by JS */}
-      <div
-        ref={thumbRef}
-        className="absolute right-0 bg-[#0a3667]"
-        style={{
-          marginLeft: "2px",
-          marginRight: "2px",
-          width: "calc(100% - 4px)",
-          display: "none", // Initially hidden
-        }}
-      />
-    </div>
-  );
-}
-
-// Helper component: MenuListItem
-function MenuListItem({
-  text,
-  isSelected,
-  onClick,
-  backlightOn = true,
-  showChevron = true,
-  value,
-}: {
-  text: string;
-  isSelected: boolean;
-  onClick: () => void;
-  backlightOn?: boolean;
-  showChevron?: boolean;
-  value?: string;
-}) {
-  return (
-    <div
-      onClick={onClick}
-      className={cn(
-        "pl-2 cursor-pointer font-chicago text-[16px] flex justify-between items-center",
-        showChevron || value ? "pr-4" : "pr-2", // Always use extra padding for items with chevron or value
-        isSelected
-          ? backlightOn
-            ? "bg-[#0a3667] text-[#c5e0f5] [text-shadow:1px_1px_0_rgba(0,0,0,0.15)]"
-            : "bg-[#0a3667] text-[#8a9da9] [text-shadow:1px_1px_0_rgba(0,0,0,0.15)]"
-          : "text-[#0a3667] hover:bg-[#c0d8f0] [text-shadow:1px_1px_0_rgba(0,0,0,0.15)]"
-      )}
-    >
-      <span className="whitespace-nowrap overflow-hidden text-ellipsis flex-1 mr-2">
-        {text}
-      </span>
-      {value ? (
-        <span className="flex-shrink-0">{value}</span>
-      ) : (
-        showChevron && <span className="flex-shrink-0">{">"}</span>
-      )}
-    </div>
-  );
-}
-
-// Helper component: ScrollingText
-function ScrollingText({
-  text,
-  className,
-  isPlaying = true,
-}: {
-  text: string;
-  className?: string;
-  isPlaying?: boolean;
-}) {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const textRef = useRef<HTMLDivElement>(null);
-  const [shouldScroll, setShouldScroll] = useState(false);
-  const [contentWidth, setContentWidth] = useState(0);
-  const paddingWidth = 20; // Width of padding between text duplicates
-
-  // Check if text needs to scroll (is wider than container)
-  useEffect(() => {
-    if (containerRef.current && textRef.current) {
-      const newContainerWidth = containerRef.current.clientWidth;
-      const newContentWidth = textRef.current.scrollWidth;
-
-      setContentWidth(newContentWidth);
-      setShouldScroll(newContentWidth > newContainerWidth);
-    }
-  }, [text, containerRef, textRef]);
-
-  return (
-    <div
-      ref={containerRef}
-      className={cn(
-        "relative overflow-hidden",
-        !shouldScroll && "flex justify-center",
-        className
-      )}
-    >
-      {shouldScroll ? (
-        <div className="inline-block whitespace-nowrap">
-          <motion.div
-            animate={{
-              x: isPlaying ? [0, -(contentWidth + paddingWidth)] : 0,
-            }}
-            transition={
-              isPlaying
-                ? {
-                    duration: Math.max(text.length * 0.15, 8),
-                    ease: "linear",
-                    repeat: Infinity,
-                  }
-                : {
-                    duration: 0.3,
-                  }
-            }
-            style={{ display: "inline-flex" }}
-          >
-            <span ref={textRef} style={{ paddingRight: `${paddingWidth}px` }}>
-              {text}
-            </span>
-            <span style={{ paddingRight: `${paddingWidth}px` }} aria-hidden>
-              {text}
-            </span>
-          </motion.div>
-        </div>
-      ) : (
-        <div ref={textRef} className="whitespace-nowrap text-center">
-          {text}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Helper component: StatusDisplay
-function StatusDisplay({ message }: { message: string }) {
-  return (
-    <div className="absolute top-4 left-4 pointer-events-none">
-      <div className="relative">
-        <div className="font-chicago text-white text-xl relative z-10">
-          {message}
-        </div>
-        <div
-          className="font-chicago text-black text-xl absolute inset-0"
-          style={{
-            WebkitTextStroke: "3px black",
-            textShadow: "none",
-          }}
-        >
-          {message}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Define the props interface for IpodScreen
-interface IpodScreenProps {
-  currentTrack: Track | null;
-  isPlaying: boolean;
-  elapsedTime: number;
-  totalTime: number;
-  menuMode: boolean;
-  menuHistory: {
-    title: string;
-    items: {
-      label: string;
-      action: () => void;
-      showChevron?: boolean;
-      value?: string;
-    }[];
-    selectedIndex: number;
-  }[];
-  selectedMenuItem: number;
-  onSelectMenuItem: (index: number) => void;
-  currentIndex: number;
-  tracksLength: number;
-  backlightOn: boolean;
-  menuDirection: "forward" | "backward";
-  onMenuItemAction: (action: () => void) => void;
-  showVideo: boolean;
-  playerRef: React.RefObject<ReactPlayer | null>;
-  handleTrackEnd: () => void;
-  handleProgress: (state: { playedSeconds: number }) => void;
-  handleDuration: (duration: number) => void;
-  handlePlay: () => void;
-  handlePause: () => void;
-  handleReady: () => void;
-  loopCurrent: boolean;
-  statusMessage: string | null;
-  onToggleVideo: () => void;
-  lcdFilterOn: boolean;
-  ipodVolume: number;
-  showStatusCallback: (message: string) => void;
-  showLyrics: boolean;
-  lyricsAlignment: LyricsAlignment;
-  chineseVariant: ChineseVariant;
-  koreanDisplay: KoreanDisplay;
-  japaneseFurigana: JapaneseFurigana;
-  lyricOffset: number;
-  adjustLyricOffset: (deltaMs: number) => void;
-  registerActivity: () => void;
-  isFullScreen: boolean;
-  lyricsControls: ReturnType<typeof useLyrics>;
-}
-
-// Main IpodScreen component
 export function IpodScreen({
   currentTrack,
   isPlaying,
@@ -419,18 +68,6 @@ export function IpodScreen({
   lyricsControls,
 }: IpodScreenProps) {
   const { t } = useTranslation();
-  // Animation variants for menu transitions
-  const menuVariants = {
-    enter: (direction: "forward" | "backward") => ({
-      x: direction === "forward" ? "100%" : "-100%",
-    }),
-    center: {
-      x: 0,
-    },
-    exit: (direction: "forward" | "backward") => ({
-      x: direction === "forward" ? "-100%" : "100%",
-    }),
-  };
 
   // Current menu title
   const currentMenuTitle = menuMode
@@ -442,8 +79,6 @@ export function IpodScreen({
   // Refs
   const menuScrollRef = useRef<HTMLDivElement>(null);
   const menuItemsRef = useRef<(HTMLDivElement | null)[]>([]);
-
-  // Need scroll flag
   const needScrollRef = useRef(false);
 
   const masterVolume = useAppStore((s) => s.masterVolume);
@@ -454,54 +89,41 @@ export function IpodScreen({
     menuItemsRef.current = Array(count).fill(null);
   };
 
-  // More direct scroll approach that doesn't rely on refs being attached yet
+  // Scroll to selected item
   const forceScrollToSelected = () => {
-    // Return if we're not in menu mode
     if (!menuMode || menuHistory.length === 0) return;
 
-    // Get the current menu's container
     const container = document.querySelector(
       ".ipod-menu-container"
     ) as HTMLElement;
     if (!container) return;
 
-    // Get all menu items
     const menuItems = Array.from(container.querySelectorAll(".ipod-menu-item"));
     if (!menuItems.length) return;
 
-    // Exit if selectedMenuItem is out of bounds
     if (selectedMenuItem < 0 || selectedMenuItem >= menuItems.length) return;
 
-    // Get the selected item
     const selectedItem = menuItems[selectedMenuItem] as HTMLElement;
     if (!selectedItem) return;
 
-    // Calculate scroll position
     const containerHeight = container.clientHeight;
     const itemTop = selectedItem.offsetTop;
     const itemHeight = selectedItem.offsetHeight;
     const scrollTop = container.scrollTop;
-
-    // Use smooth scrolling with a small buffer to prevent edge flickering
-    // Add a 2px buffer at top and bottom to prevent edge flickering
     const buffer = 2;
 
-    // If item is below the visible area
     if (itemTop + itemHeight > scrollTop + containerHeight - buffer) {
       container.scrollTo({
         top: itemTop + itemHeight - containerHeight + buffer,
         behavior: "instant" as ScrollBehavior,
       });
-    }
-    // If item is above the visible area
-    else if (itemTop < scrollTop + buffer) {
+    } else if (itemTop < scrollTop + buffer) {
       container.scrollTo({
         top: Math.max(0, itemTop - buffer),
         behavior: "instant" as ScrollBehavior,
       });
     }
 
-    // Force scroll to top for first item
     if (selectedMenuItem === 0) {
       container.scrollTo({
         top: 0,
@@ -509,7 +131,6 @@ export function IpodScreen({
       });
     }
 
-    // For last item, ensure it's fully visible
     if (selectedMenuItem === menuItems.length - 1) {
       container.scrollTo({
         top: Math.max(0, itemTop - (containerHeight - itemHeight) + buffer),
@@ -517,22 +138,16 @@ export function IpodScreen({
       });
     }
 
-    // Reset need scroll flag
     needScrollRef.current = false;
   };
 
   // Trigger scroll on various conditions
   useEffect(() => {
     if (menuMode && menuHistory.length > 0) {
-      // Flag that we need to scroll
       needScrollRef.current = true;
-
-      // Try immediately (in case DOM is ready)
       forceScrollToSelected();
 
-      // Schedule multiple attempts with increasing delays
       const attempts = [50, 100, 250, 500, 1000];
-
       attempts.forEach((delay) => {
         setTimeout(() => {
           if (needScrollRef.current) {
@@ -561,28 +176,27 @@ export function IpodScreen({
         backlightOn
           ? "bg-[#c5e0f5] bg-gradient-to-b from-[#d1e8fa] to-[#e0f0fc]"
           : "bg-[#8a9da9] contrast-65 saturate-50",
-        // Add the soft blue glow when both LCD filter and backlight are on
         lcdFilterOn &&
           backlightOn &&
           "shadow-[0_0_10px_2px_rgba(197,224,245,0.05)]"
       )}
       style={{
-        minWidth: '100%',
-        minHeight: '150px',
-        maxWidth: '100%',
-        maxHeight: '150px',
-        position: 'relative',
-        contain: 'layout style paint',
-        WebkitUserSelect: 'none',
-        WebkitTouchCallout: 'none',
+        minWidth: "100%",
+        minHeight: "150px",
+        maxWidth: "100%",
+        maxHeight: "150px",
+        position: "relative",
+        contain: "layout style paint",
+        WebkitUserSelect: "none",
+        WebkitTouchCallout: "none",
       }}
     >
-      {/* LCD screen overlay with scan lines - only show when LCD filter is on */}
+      {/* LCD screen overlay with scan lines */}
       {lcdFilterOn && (
         <div className="absolute inset-0 pointer-events-none z-25 lcd-scan-lines"></div>
       )}
 
-      {/* Glass reflection effect - only show when LCD filter is on */}
+      {/* Glass reflection effect */}
       {lcdFilterOn && (
         <div className="absolute inset-0 pointer-events-none z-25 lcd-reflection"></div>
       )}
@@ -601,23 +215,18 @@ export function IpodScreen({
           <div
             className="w-full h-[calc(100%+120px)] mt-[-60px]"
             onClick={(e) => {
-              // Ensure taps on the lyrics overlay also toggle play / video as expected
               e.stopPropagation();
               registerActivity();
               if (!isPlaying) {
-                // When not playing, show video and resume playback
                 if (!showVideo) {
                   onToggleVideo();
-                  // Give React a moment to render the player before playing
                   setTimeout(() => {
                     handlePlay();
                   }, 100);
                 } else {
-                  // Video already visible; just resume playback
                   handlePlay();
                 }
               } else {
-                // Playing → hide video
                 onToggleVideo();
               }
             }}
@@ -649,10 +258,8 @@ export function IpodScreen({
                     disablekb: 1,
                     playsinline: 1,
                     enablejsapi: 1,
-                    // Origin for YouTube postMessage communication
                     origin: window.location.origin,
                   },
-                  // Required for Tauri: sets referrer policy on iframe to prevent YouTube Error 153
                   embedOptions: {
                     referrerPolicy: "strict-origin-when-cross-origin",
                   },
@@ -669,7 +276,6 @@ export function IpodScreen({
                 className="absolute inset-0 z-30"
                 onClick={(e) => {
                   e.stopPropagation();
-                  // Only resume playback; keep video visible
                   if (!isPlaying) {
                     handlePlay();
                   } else {
@@ -693,7 +299,7 @@ export function IpodScreen({
               )}
             </AnimatePresence>
 
-            {/* Activity Indicator - top right, aligned with status display */}
+            {/* Activity Indicator */}
             <AnimatePresence>
               {(lyricsControls.isLoading || lyricsControls.isTranslating) && (
                 <motion.div
@@ -732,7 +338,6 @@ export function IpodScreen({
                 showStatusCallback(
                   `${t("apps.ipod.status.offset")} ${sign}${(newOffset / 1000).toFixed(2)}s`
                 );
-                // Force immediate update of lyrics display with new offset
                 const updatedTime = elapsedTime + newOffset / 1000;
                 lyricsControls.updateCurrentTimeManually(updatedTime);
               }}
@@ -741,7 +346,7 @@ export function IpodScreen({
         </div>
       )}
 
-      {/* Title bar - not animated, immediately swaps */}
+      {/* Title bar */}
       <div className="border-b border-[#0a3667] py-0 px-2 font-chicago text-[16px] flex items-center sticky top-0 z-10 text-[#0a3667] [text-shadow:1px_1px_0_rgba(0,0,0,0.15)]">
         <div
           className={`w-6 flex items-center justify-start font-chicago ${
@@ -758,7 +363,7 @@ export function IpodScreen({
         </div>
       </div>
 
-      {/* Content area - this animates/slides */}
+      {/* Content area */}
       <div className="relative h-[calc(100%-26px)]">
         <AnimatePresence initial={false} custom={menuDirection} mode="sync">
           {menuMode ? (
@@ -772,7 +377,6 @@ export function IpodScreen({
               transition={{ duration: 0.2, ease: "easeInOut" }}
               custom={menuDirection}
               onAnimationComplete={() => {
-                // Flag that we need to scroll and trigger the scroll logic
                 needScrollRef.current = true;
                 forceScrollToSelected();
               }}
@@ -830,21 +434,15 @@ export function IpodScreen({
                 if (!menuMode && currentTrack) {
                   registerActivity();
                   if (!isPlaying) {
-                    // If the video is currently hidden, show it first so the
-                    // user gesture also applies to the player element.
                     if (!showVideo) {
                       onToggleVideo();
-                      // Give React a moment to render the player before
-                      // resuming playback so the gesture is linked.
                       setTimeout(() => {
                         handlePlay();
                       }, 100);
                     } else {
-                      // Resume playback immediately when video already visible
                       handlePlay();
                     }
                   } else {
-                    // Already playing — simply toggle video visibility
                     onToggleVideo();
                   }
                 }
