+++ src/apps/ipod/components/screen/index.ts
@@ -1,0 +1,6 @@
+// Re-export all screen components
+export { BatteryIndicator } from "./BatteryIndicator";
+export { Scrollbar } from "./Scrollbar";
+export { MenuListItem } from "./MenuListItem";
+export { ScrollingText } from "./ScrollingText";
+export { StatusDisplay } from "./StatusDisplay";
