+++ src/apps/ipod/hooks/index.ts
@@ -1,0 +1,2 @@
+// Re-export all iPod hooks
+export { useLibraryUpdateChecker } from "./useLibraryUpdateChecker";
